[chore]클린 아키텍처 구조 생성(Facade Service, Domain Service는 추후 필요 시 생성)
