+++ src/main/java/kr/hhplus/be/server/application/balance/BalanceRepository.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.application.balance;
+
+import kr.hhplus.be.server.domain.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BalanceRepository extends JpaRepository<User, Long> {
+
+}
