+++ src/main/java/kr/hhplus/be/server/application/balance/BalanceSerivce.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.application.balance;
+
+public interface BalanceSerivce {
+
+}
