+++ src/main/java/kr/hhplus/be/server/application/balance/BalanceUsecase.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.balance;
+
+public interface BalanceUsecase {
+}
