+++ src/main/java/kr/hhplus/be/server/application/balance/impl/BalanceServiceImpl.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.balance.impl;
+
+import kr.hhplus.be.server.application.balance.BalanceSerivce;
+import org.springframework.stereotype.Service;
+
+@Service
+public class BalanceServiceImpl implements BalanceSerivce {
+}
