+++ src/main/java/kr/hhplus/be/server/application/coupon/CouponRepository.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.application.coupon;
+
+import kr.hhplus.be.server.domain.coupon.Coupon;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CouponRepository extends JpaRepository<Coupon, Long> {
+
+}
