+++ src/main/java/kr/hhplus/be/server/application/coupon/CouponService.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.application.coupon;
+
+public interface CouponService {
+
+}
