+++ src/main/java/kr/hhplus/be/server/application/coupon/CouponUsecase.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.coupon;
+
+public interface CouponUsecase {
+}
