+++ src/main/java/kr/hhplus/be/server/application/coupon/impl/CouponServiceImpl.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.coupon.impl;
+
+import kr.hhplus.be.server.application.coupon.CouponService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CouponServiceImpl implements CouponService {
+}
