+++ src/main/java/kr/hhplus/be/server/application/order/OrderRepository.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.application.order;
+
+import kr.hhplus.be.server.domain.order.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
+}
