+++ src/main/java/kr/hhplus/be/server/application/order/OrderService.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.order;
+
+public interface OrderService {
+}
