+++ src/main/java/kr/hhplus/be/server/application/order/OrderUsecase.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.order;
+
+public interface OrderUsecase {
+}
