+++ src/main/java/kr/hhplus/be/server/application/order/impl/OrderServiceImpl.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.order.impl;
+
+import kr.hhplus.be.server.application.order.OrderService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class OrderServiceImpl implements OrderService {
+}
