+++ src/main/java/kr/hhplus/be/server/application/payment/PaymentRepository.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.application.payment;
+
+import kr.hhplus.be.server.domain.payment.Payment;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PaymentRepository extends JpaRepository<Payment, Long> {
+
+}
