+++ src/main/java/kr/hhplus/be/server/application/payment/PaymentService.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.payment;
+
+public interface PaymentService {
+}
