+++ src/main/java/kr/hhplus/be/server/application/payment/PaymentUsecase.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.payment;
+
+public interface PaymentUsecase {
+}
