+++ src/main/java/kr/hhplus/be/server/application/payment/impl/PaymentServiceImpl.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.payment.impl;
+
+import kr.hhplus.be.server.application.payment.PaymentService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class PaymentServiceImpl implements PaymentService {
+}
