+++ src/main/java/kr/hhplus/be/server/application/product/ProductRepository.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.application.product;
+
+import kr.hhplus.be.server.domain.product.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+}
