+++ src/main/java/kr/hhplus/be/server/application/product/ProductService.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.application.product;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public interface ProductService {
+}
