+++ src/main/java/kr/hhplus/be/server/application/product/ProductUsecase.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.product;
+
+public interface ProductUsecase {
+}
