+++ src/main/java/kr/hhplus/be/server/application/product/impl/ProductServiceImpl.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.product.impl;
+
+import kr.hhplus.be.server.application.product.ProductService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProductServiceImpl implements ProductService {
+}
