+++ src/main/java/kr/hhplus/be/server/application/user/UserRepository.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.user;
+
+import kr.hhplus.be.server.domain.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User,Long> {
+
+}
