+++ src/main/java/kr/hhplus/be/server/application/user/UserUsecase.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.application.user;
+
+public interface UserUsecase {
+}
