--- src/main/java/kr/hhplus/be/server/infrastructure/order/OrderRepository.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.infrastructure.order;
-
-import kr.hhplus.be.server.domain.order.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
-
-}
