--- src/main/java/kr/hhplus/be/server/infrastructure/payment/PaymentRepository.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.infrastructure.payment;
-
-import kr.hhplus.be.server.domain.payment.Payment;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
-
-}
