--- src/main/java/kr/hhplus/be/server/infrastructure/product/ProductRepository.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.infrastructure.product;
-
-import kr.hhplus.be.server.domain.product.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-}
