--- src/main/java/kr/hhplus/be/server/infrastructure/user/UserRepository.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.infrastructure.user;
-
-import kr.hhplus.be.server.domain.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User,Long> {
-
-}
