+++ src/main/java/kr/hhplus/be/server/persistence/balance/BalanceRepositoryImpl.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.persistence.balance;
+
+public class BalanceRepositoryImpl {
+}
