+++ src/main/java/kr/hhplus/be/server/persistence/coupon/CouponIssuedInfoRepositoryImpl.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.persistence.coupon;
+
+public class CouponIssuedInfoRepositoryImpl {
+}
