+++ src/main/java/kr/hhplus/be/server/persistence/order/OrderRepositoryImpl.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.persistence.order;
+
+public class OrderRepositoryImpl {
+}
