+++ src/main/java/kr/hhplus/be/server/persistence/payment/PaymentRepositoryImpl.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.persistence.payment;
+
+public class PaymentRepositoryImpl {
+}
