+++ src/main/java/kr/hhplus/be/server/persistence/product/ProductRepositoryImpl.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.persistence.product;
+
+public class ProductRepositoryImpl {
+}
