+++ src/main/java/kr/hhplus/be/server/persistence/user/UserRepositoryImpl.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.persistence.user;
+
+public class UserRepositoryImpl {
+}
