--- src/main/java/kr/hhplus/be/server/service/balance/BalanceSerivce.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.service.balance;
-
-public interface BalanceSerivce {
-
-}
