--- src/main/java/kr/hhplus/be/server/service/balance/impl/BalanceServiceImpl.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.service.balance.impl;
-
-import kr.hhplus.be.server.service.balance.BalanceSerivce;
-import org.springframework.stereotype.Service;
-
-@Service
-public class BalanceServiceImpl implements BalanceSerivce {
-}
