--- src/main/java/kr/hhplus/be/server/service/coupon/CouponService.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.service.coupon;
-
-public interface CouponService {
-
-}
