--- src/main/java/kr/hhplus/be/server/service/coupon/impl/CouponServiceImpl.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.service.coupon.impl;
-
-import kr.hhplus.be.server.service.coupon.CouponService;
-import org.springframework.stereotype.Service;
-
-@Service
-public class CouponServiceImpl implements CouponService {
-}
