--- src/main/java/kr/hhplus/be/server/service/order/OrderService.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.service.order;
-
-public interface OrderService {
-}
