--- src/main/java/kr/hhplus/be/server/service/order/impl/OrderServiceImpl.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.service.order.impl;
-
-import kr.hhplus.be.server.service.order.OrderService;
-import org.springframework.stereotype.Service;
-
-@Service
-public class OrderServiceImpl implements OrderService {
-}
