--- src/main/java/kr/hhplus/be/server/service/payment/PaymentService.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.service.payment;
-
-public interface PaymentService {
-}
