--- src/main/java/kr/hhplus/be/server/service/payment/impl/PaymentServiceImpl.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.service.payment.impl;
-
-import kr.hhplus.be.server.service.payment.PaymentService;
-import org.springframework.stereotype.Service;
-
-@Service
-public class PaymentServiceImpl implements PaymentService {
-}
