--- src/main/java/kr/hhplus/be/server/service/product/ProductService.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.service.product;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface ProductService {
-}
