--- src/main/java/kr/hhplus/be/server/service/product/impl/ProductServiceImpl.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.service.product.impl;
-
-import kr.hhplus.be.server.service.product.ProductService;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ProductServiceImpl implements ProductService {
-}
